bcc_zip_archive_open should not loop indefinitely for bad archives

+++ src/cc/bcc_zip.c
@@ -202,9 +202,9 @@ static int find_central_directory(struct bcc_zip_archive* archive) {
   // Because the end of central directory ends with a variable length array of
   // up to 0xFFFF bytes we can't know exactly where it starts and need to
   // search for it at the end of the file, scanning the (limit, offset] range.
-  uint32_t offset =
-      archive->size - sizeof(struct end_of_central_directory_record);
-  int64_t limit = (int64_t)offset - (1 << 16);
+  int64_t offset =
+      (int64_t)archive->size - sizeof(struct end_of_central_directory_record);
+  int64_t limit = offset - (1 << 16);
   for (; offset >= 0 && offset > limit && rc == -1; offset--) {
     rc = try_parse_end_of_central_directory(archive, offset);
   }
